+++ packages/env/lib/commands/run.js
@@ -18,14 +18,14 @@ const initConfig = require( '../init-config' );
  * @param {Object}   options
  * @param {string}   options.container The Docker container to run the command on.
  * @param {string[]} options.command   The command to run.
- * @param {string}   options.cwd       The working directory for the command to be executed from.
+ * @param {string}   options.envCwd    The working directory for the command to be executed from.
  * @param {Object}   options.spinner   A CLI spinner which indicates progress.
  * @param {boolean}  options.debug     True if debug mode is enabled.
  */
 module.exports = async function run( {
 	container,
 	command,
-	cwd,
+	envCwd,
 	spinner,
 	debug,
 } ) {
@@ -36,7 +36,7 @@ module.exports = async function run( {
 	// Shows a contextual tip for the given command.
 	showCommandTips( command, container, spinner );
 
-	await spawnCommandDirectly( config, container, command, cwd, spinner );
+	await spawnCommandDirectly( config, container, command, envCwd, spinner );
 
 	spinner.text = `Ran \`${ command }\` in '${ container }'.`;
 };
@@ -47,22 +47,20 @@ module.exports = async function run( {
  * @param {WPConfig} config    The wp-env configuration.
  * @param {string}   container The Docker container to run the command on.
  * @param {string}   command   The command to run.
- * @param {string}   cwd       The working directory for the command to be executed from.
+ * @param {string}   envCwd    The working directory for the command to be executed from.
  * @param {Object}   spinner   A CLI spinner which indicates progress.
  */
-function spawnCommandDirectly( config, container, command, cwd, spinner ) {
+function spawnCommandDirectly( config, container, command, envCwd, spinner ) {
 	const composeCommand = [
 		'-f',
 		config.dockerComposeConfigPath,
 		'run',
 		'-w',
-		cwd,
+		envCwd,
 		'--rm',
 		container,
 		...command.split( ' ' ), // The command will fail if passed as a complete string.
 	];
-
-	console.log( composeCommand );
 
 	return new Promise( ( resolve, reject ) => {
 		// Note: since the npm docker-compose package uses the -T option, we